MISC: move AyaDocile to base
