+++ base/src/main/java/org/aya/concrete/Expr.java
@@ -11,6 +11,7 @@ import org.aya.concrete.visitor.ExprView;
 import org.aya.core.pat.Pat;
 import org.aya.distill.BaseDistiller;
 import org.aya.distill.ConcreteDistiller;
+import org.aya.generic.AyaDocile;
 import org.aya.generic.ParamLike;
 import org.aya.pretty.doc.Doc;
 import org.aya.ref.LocalVar;
@@ -20,7 +21,6 @@ import org.aya.resolve.visitor.ExprResolver;
 import org.aya.resolve.visitor.StmtShallowResolver;
 import org.aya.tyck.ExprTycker;
 import org.aya.util.binop.BinOpParser;
-import org.aya.util.distill.AyaDocile;
 import org.aya.util.distill.DistillerOptions;
 import org.aya.util.error.SourceNode;
 import org.aya.util.error.SourcePos;
