+++ base/src/main/java/org/aya/concrete/Pattern.java
@@ -6,11 +6,11 @@ import kala.collection.immutable.ImmutableSeq;
 import kala.control.Option;
 import org.aya.distill.BaseDistiller;
 import org.aya.distill.ConcreteDistiller;
+import org.aya.generic.AyaDocile;
 import org.aya.pretty.doc.Doc;
 import org.aya.ref.LocalVar;
 import org.aya.ref.Var;
 import org.aya.util.binop.BinOpParser;
-import org.aya.util.distill.AyaDocile;
 import org.aya.util.distill.DistillerOptions;
 import org.aya.util.error.SourceNode;
 import org.aya.util.error.SourcePos;
