+++ base/src/main/java/org/aya/concrete/stmt/Stmt.java
@@ -7,13 +7,13 @@ import kala.tuple.Unit;
 import org.aya.concrete.desugar.Desugarer;
 import org.aya.concrete.remark.Remark;
 import org.aya.distill.ConcreteDistiller;
+import org.aya.generic.AyaDocile;
 import org.aya.pretty.doc.Doc;
 import org.aya.resolve.ResolveInfo;
 import org.aya.resolve.module.ModuleLoader;
 import org.aya.resolve.visitor.StmtResolver;
 import org.aya.resolve.visitor.StmtShallowResolver;
 import org.aya.tyck.order.TyckUnit;
-import org.aya.util.distill.AyaDocile;
 import org.aya.util.distill.DistillerOptions;
 import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
