+++ base/src/main/java/org/aya/core/Matching.java
@@ -5,8 +5,8 @@ package org.aya.core;
 import kala.collection.immutable.ImmutableSeq;
 import org.aya.core.pat.Pat;
 import org.aya.core.term.Term;
+import org.aya.generic.AyaDocile;
 import org.aya.pretty.doc.Doc;
-import org.aya.util.distill.AyaDocile;
 import org.aya.util.distill.DistillerOptions;
 import org.aya.util.error.SourcePos;
 import org.jetbrains.annotations.NotNull;
