+++ base/src/main/java/org/aya/core/visitor/Subst.java
@@ -8,9 +8,9 @@ import kala.collection.mutable.MutableMap;
 import kala.collection.mutable.MutableTreeMap;
 import org.aya.core.term.Term;
 import org.aya.distill.BaseDistiller;
+import org.aya.generic.AyaDocile;
 import org.aya.pretty.doc.Doc;
 import org.aya.ref.Var;
-import org.aya.util.distill.AyaDocile;
 import org.aya.util.distill.DistillerOptions;
 import org.jetbrains.annotations.Debug;
 import org.jetbrains.annotations.NotNull;
