+++ base/src/main/java/org/aya/tyck/TyckState.java
@@ -13,13 +13,13 @@ import org.aya.core.term.CallTerm;
 import org.aya.core.term.Term;
 import org.aya.core.visitor.TermConsumer;
 import org.aya.core.visitor.VarConsumer;
+import org.aya.generic.AyaDocile;
 import org.aya.pretty.doc.Doc;
 import org.aya.tyck.env.LocalCtx;
 import org.aya.tyck.error.HoleProblem;
 import org.aya.tyck.trace.Trace;
 import org.aya.tyck.unify.DefEq;
 import org.aya.util.Ordering;
-import org.aya.util.distill.AyaDocile;
 import org.aya.util.distill.DistillerOptions;
 import org.aya.util.error.SourcePos;
 import org.aya.util.error.WithPos;
