+++ base/src/main/java/org/aya/tyck/error/BadTypeError.java
@@ -7,13 +7,13 @@ import org.aya.concrete.Expr;
 import org.aya.concrete.stmt.QualifiedID;
 import org.aya.core.def.FieldDef;
 import org.aya.core.term.Term;
+import org.aya.generic.AyaDocile;
 import org.aya.generic.ExprProblem;
 import org.aya.generic.util.NormalizeMode;
 import org.aya.pretty.doc.Doc;
 import org.aya.pretty.doc.Style;
 import org.aya.ref.DefVar;
 import org.aya.tyck.TyckState;
-import org.aya.util.distill.AyaDocile;
 import org.aya.util.distill.DistillerOptions;
 import org.aya.util.error.SourcePos;
 import org.jetbrains.annotations.NotNull;
