+++ cli/src/main/java/org/aya/cli/repl/AyaRepl.java
@@ -7,11 +7,11 @@ import org.aya.cli.repl.jline.AyaCompleters;
 import org.aya.cli.repl.jline.JlineRepl;
 import org.aya.cli.single.CliReporter;
 import org.aya.cli.utils.MainArgs;
+import org.aya.generic.AyaDocile;
 import org.aya.generic.util.NormalizeMode;
 import org.aya.prelude.GeneratedVersion;
 import org.aya.pretty.doc.Doc;
 import org.aya.repl.*;
-import org.aya.util.distill.AyaDocile;
 import org.aya.util.distill.DistillerOptions;
 import org.aya.util.reporter.Problem;
 import org.jetbrains.annotations.NotNull;
