+++ cli/src/main/java/org/aya/cli/single/SingleFileCompiler.java
@@ -11,6 +11,7 @@ import org.aya.concrete.stmt.Decl;
 import org.aya.core.def.Def;
 import org.aya.core.def.PrimDef;
 import org.aya.core.serde.Serializer;
+import org.aya.generic.AyaDocile;
 import org.aya.pretty.backend.string.StringPrinterConfig;
 import org.aya.pretty.doc.Doc;
 import org.aya.resolve.ModuleCallback;
@@ -20,7 +21,6 @@ import org.aya.resolve.module.CachedModuleLoader;
 import org.aya.resolve.module.FileModuleLoader;
 import org.aya.resolve.module.ModuleListLoader;
 import org.aya.tyck.trace.Trace;
-import org.aya.util.distill.AyaDocile;
 import org.aya.util.distill.DistillerOptions;
 import org.aya.util.error.SourceFileLocator;
 import org.aya.util.reporter.CountingReporter;
